+++ doc/changelog.rst
@@ -3,8 +3,15 @@
 Changelog
 =========
 
-3.0.1 (unreleased)
+3.1.0 (unreleased)
 ------------------
+
+Changes
+~~~~~~~
+
+- heyoka.py now requires version 3.1.0 of the
+  heyoka C++ library
+  (`#140 <https://github.com/bluescarni/heyoka.py/pull/140>`__).
 
 Fix
 ~~~